% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsets.R
\name{calc_offsets}
\alias{calc_offsets}
\title{Calculate offsets}
\usage{
calc_offsets(segments, segment_id, segment_min = "min", segment_max = "max")
}
\arguments{
\item{segments}{A data.frame describing max and min of each segment}

\item{segment_id}{A variable to describe the segments for which to calculate offsets. Converted to factor. Need to be the same as the number of rows.}

\item{segment_min}{A variable in offsets describing the minimum of each segment}

\item{segment_max}{A variable in offsets describing the maximum of each segment}
}
\value{
A data.frame
}
\description{
Calculate offsets
}
\examples{

segments <- data.frame(
  segment = LETTERS[1:4],
  min = c(0, 2, 0, 0),
  max = c(3,4,3,2)
)
calc_offsets(segments, "segment", "min", "max")
}
