% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsets.R
\name{offset_data}
\alias{offset_data}
\title{Offset continuous data by segment}
\usage{
offset_data(data, segments, segment_id, disp, segment_min = "min",
  segment_max = "max")
}
\arguments{
\item{data}{A data.frame for which to calculate offsets}

\item{segments}{A data.frame describing max and min of each segment}

\item{segment_id}{A commmon variable identifying segments between data and offsets.}

\item{disp}{A continuous variable in data for which offsets will be added}

\item{segment_min}{A variable in offsets describing the minimum of each segment}

\item{segment_max}{A variable in offsets describing the maximum of each segment}
}
\value{
A data.frame
}
\description{
Offset continuous data by segment
}
\examples{

segments <- data.frame(
  segment = LETTERS[1:4],
  min = c(0, 2, 0, 0),
  max = c(3,4,3,2)
)
dat <- data.frame(
  segment = c("A", "C", "D" ),
  dist = c(1, 2, 1))
  offset_data(dat, segments, "segment", "dist", "min", "max")
}
